+++ src/lib/definitions/generator-configuration.ts
@@ -44,7 +44,10 @@ export const CharacteristicDeprecatedNames: Map<string, string> = new Map([ // k
 ]);
 
 export const CharacteristicValidValuesOverride: Map<string, Record<string, string>> = new Map([
+  ["camera-operating-mode-indicator", { "0": "Disable", "1": "Enable" }],
   ["closed-captions", { "0": "Disabled", "1": "Enabled" }],
+  ["event-snapshots-active", { "0": "Disable", "1": "Enable" }],
+  ["homekit-camera-active", { "0": "Off", "1": "On" }],
   ["input-device-type", { "0": "Other", "1": "TV", "2": "Recording", "3": "Tuner", "4": "Playback", "5": "Audio System" }],
   ["input-source-type", { "0": "Other", "1": "Home Screen", "2": "Tuner", "3": "HDMI", "4": "Composite Video", "5": "S Video",
     "6": "Component Video", "7": "DVI", "8": "AirPlay", "9": "USB", "10": "Application" }],
@@ -52,6 +55,7 @@ export const CharacteristicValidValuesOverride: Map<string, Record<string, strin
   ["manually-disabled", { "0": "Enabled", "1": "Disabled" }],
   ["media-state.current", { "0": "Play", "1": "Pause", "2": "Stop", "4": "LOADING", "5": "Interrupted" }],
   ["media-state.target", { "0": "Play", "1": "Pause", "2": "Stop" }],
+  ["periodic-snapshots-active", { "0": "Disable", "1": "Enable" }],
   ["picture-mode", { "0": "Other", "1": "Standard", "2": "Calibrated", "3": "Calibrated Dark", "4": "Vivid", "5": "Game", "6": "Computer", "7": "Custom" }],
   ["power-mode-selection", { "0": "Show", "1": "Hide" }],
   ["recording-audio-active", { "0": "Disable", "1": "Enable" }],
@@ -60,6 +64,7 @@ export const CharacteristicValidValuesOverride: Map<string, Record<string, strin
   ["router-status", { "0": "Ready", "1": "Not Ready" }],
   ["siri-input-type", { "0": "Push Button Triggered Apple TV" }],
   ["sleep-discovery-mode", { "0": "Not Discoverable", "1": "Always Discoverable" }],
+  ["third-party-camera-active", { "0": "Off", "1": "On" }],
   ["visibility-state.current", { "0": "Shown", "1": "Hidden" }],
   ["visibility-state.target", { "0": "Shown", "1": "Hidden" }],
   ["volume-control-type", { "0": "None", "1": "Relative", "2": "Relative With Current", "3": "Absolute" }],
@@ -67,11 +72,7 @@ export const CharacteristicValidValuesOverride: Map<string, Record<string, strin
   ["wifi-satellite-status", { "0": "Unknown", "1": "Connected", "2": "Not Connected" }],
 ] as [string, Record<string, string>][]);
 
-export const CharacteristicClassAdditions: Map<string, string[]> = new Map([
-  ["humidifier-dehumidifier.state.target", [
-    "/**\n   * @deprecated Removed in iOS 11. Use {@link HUMIDIFIER_OR_DEHUMIDIFIER} instead.\n   */\n  public static readonly AUTO = 0;",
-  ]],
-]);
+export const CharacteristicClassAdditions: Map<string, string[]> = new Map([]);
 
 export const CharacteristicOverriding: Map<string, (generated: GeneratedCharacteristic) => void> = new Map([
   ["rotation.speed", generated => {
@@ -219,148 +220,6 @@ export const CharacteristicManualAdditions: Map<string, GeneratedCharacteristic>
     properties: 2, // paired read
     maxLength: 64,
   }],
-  ["target-air-quality", { // some legacy characteristic, don't know where it comes from or where it was used
-    id: "target-air-quality",
-    UUID: "000000AE-0000-1000-8000-0026BB765291",
-    name: "Target Air Quality",
-    className: "TargetAirQuality",
-    deprecatedNotice: "Removed and not used anymore",
-
-    format: "uint8",
-    properties: 7, // read, write, notify
-    minValue: 0,
-    maxValue: 2,
-    validValues: {
-      "0": "EXCELLENT",
-      "1": "GOOD",
-      "2": "FAIR",
-    } as Record<string, string>,
-  }],
-  ["target-slat-state", { // some legacy characteristic, don't know where it comes from or where it was used
-    id: "target-slat-state",
-    UUID: "000000BE-0000-1000-8000-0026BB765291",
-    name: "Target Slat State",
-    className: "TargetSlatState",
-    deprecatedNotice: "Removed and not used anymore",
-
-    format: "uint8",
-    properties: 7, // read, write, notify
-    minValue: 0,
-    maxValue: 1,
-    validValues: {
-      "0": "MANUAL",
-      "1": "AUTO",
-    } as Record<string, string>,
-  }],
-
-  ["current-time", {
-    id: "current-time",
-    UUID: "0000009B-0000-1000-8000-0026BB765291",
-    name: "Current Time",
-    className: "CurrentTime",
-    deprecatedNotice: "Removed and not used anymore",
-
-    format: "string",
-    properties: 6, // read, write
-  }],
-  ["day-of-the-week", {
-    id: "day-of-the-week",
-    UUID: "00000098-0000-1000-8000-0026BB765291",
-    name: "Day of the Week",
-    className: "DayoftheWeek",
-    deprecatedNotice: "Removed and not used anymore",
-
-    format: "uint8",
-    properties: 6, // read, write
-    minValue: 1,
-    maxValue: 7,
-  }],
-  ["time-update", {
-    id: "time-update",
-    UUID: "0000009A-0000-1000-8000-0026BB765291",
-    name: "Time Update",
-    className: "TimeUpdate",
-    deprecatedNotice: "Removed and not used anymore",
-
-    format: "bool",
-    properties: 6, // read, write
-  }],
-
-  ["reachable", {
-    id: "reachable",
-    UUID: "00000063-0000-1000-8000-0026BB765291",
-    name: "Reachable",
-    className: "Reachable",
-    deprecatedNotice: "Removed and not used anymore",
-
-    format: "bool",
-    properties: 6, // read, write
-  }],
-  ["link-quality", {
-    id: "link-quality",
-    UUID: "0000009C-0000-1000-8000-0026BB765291",
-    name: "Link Quality",
-    className: "LinkQuality",
-    deprecatedNotice: "Removed and not used anymore",
-
-    format: "uint8",
-    properties: 3, // read, notify
-    minValue: 1,
-    maxValue: 4,
-  }],
-  ["category", {
-    id: "category",
-    UUID: "000000A3-0000-1000-8000-0026BB765291",
-    name: "Category",
-    className: "Category",
-    deprecatedNotice: "Removed and not used anymore",
-
-    format: "uint16",
-    properties: 3, // read, notify
-    minValue: 1,
-    maxValue: 16,
-  }],
-
-  ["configure-bridged-accessory-status", {
-    id: "configure-bridged-accessory-status",
-    UUID: "0000009D-0000-1000-8000-0026BB765291",
-    name: "Configure Bridged Accessory Status",
-    className: "ConfigureBridgedAccessoryStatus",
-    deprecatedNotice: "Removed and not used anymore",
-
-    format: "tlv8",
-    properties: 3, // read, notify
-  }],
-  ["configure-bridged-accessory", {
-    id: "configure-bridged-accessory",
-    UUID: "000000A0-0000-1000-8000-0026BB765291",
-    name: "Configure Bridged Accessory",
-    className: "ConfigureBridgedAccessory",
-    deprecatedNotice: "Removed and not used anymore",
-
-    format: "tlv8",
-    properties: 4,
-  }],
-  ["discover-bridged-accessories", {
-    id: "discover-bridged-accessories",
-    UUID: "0000009E-0000-1000-8000-0026BB765291",
-    name: "Discover Bridged Accessories",
-    className: "DiscoverBridgedAccessories",
-    deprecatedNotice: "Removed and not used anymore",
-
-    format: "uint8",
-    properties: 7, // read, write, notify
-  }],
-  ["discovered-bridged-accessories", {
-    id: "discovered-bridged-accessories",
-    UUID: "0000009F-0000-1000-8000-0026BB765291",
-    name: "Discovered Bridged Accessories",
-    className: "DiscoveredBridgedAccessories",
-    deprecatedNotice: "Removed and not used anymore",
-
-    format: "uint16",
-    properties: 3, // read, notify
-  }],
 ]);
 
 export const ServiceNameOverrides: Map<string, string> = new Map([
@@ -373,13 +232,7 @@ export const ServiceNameOverrides: Map<string, string> = new Map([
   ["nfc-access", "NFC Access"],
 ]);
 
-export const ServiceDeprecatedNames: Map<string, string> = new Map([
-  ["battery", "Battery Service"],
-  ["camera-recording-management", "Camera Event Recording Management"],
-  ["cloud-relay", "Relay"],
-  ["slats", "Slat"],
-  ["tunnel", "Tunneled BTLE Accessory Service"],
-]);
+export const ServiceDeprecatedNames: Map<string, string> = new Map([]);
 
 interface CharacteristicConfigurationOverride {
   addedRequired?: string[],
@@ -403,64 +256,6 @@ export const ServiceManualAdditions: Map<string, GeneratedService> = new Map([
 
     requiredCharacteristics: ["mute"],
     optionalCharacteristics: ["active", "volume"],
-  }],
-  ["camera-control", {
-    id: "camera-control",
-    UUID: "00000111-0000-1000-8000-0026BB765291",
-    name: "Camera Control",
-    className: "CameraControl",
-    deprecatedNotice: "This service has no usage anymore and will be ignored by iOS",
-
-    requiredCharacteristics: ["on"],
-    optionalCharacteristics: [
-      "horizontal-tilt.current",
-      "vertical-tilt.current",
-      "horizontal-tilt.target",
-      "vertical-tilt.target",
-      "night-vision",
-      "optical-zoom",
-      "digital-zoom",
-      "image-rotation",
-      "image-mirroring",
-      "name",
-    ],
-  }],
-  ["time-information", {
-    id: "time-information",
-    UUID: "00000099-0000-1000-8000-0026BB765291",
-    name: "Time Information",
-    className: "TimeInformation",
-    deprecatedNotice: "Removed and not used anymore",
-
-    requiredCharacteristics: ["current-time", "day-of-the-week", "time-update"],
-    optionalCharacteristics: ["name"],
-  }],
-
-  ["bridging-state", {
-    id: "bridging-state",
-    UUID: "00000062-0000-1000-8000-0026BB765291",
-    name: "Bridging State",
-    className: "BridgingState",
-    deprecatedNotice: "Removed and not used anymore",
-
-    requiredCharacteristics: ["reachable", "link-quality", "accessory.identifier", "category"],
-    optionalCharacteristics: ["name"],
-  }],
-
-  ["bridge-configuration", {
-    id: "bridge-configuration",
-    UUID: "000000A1-0000-1000-8000-0026BB765291",
-    name: "Bridge Configuration",
-    className: "BridgeConfiguration",
-    deprecatedNotice: "Removed and not used anymore",
-
-    requiredCharacteristics: [
-      "configure-bridged-accessory-status",
-      "discover-bridged-accessories",
-      "discovered-bridged-accessories",
-      "configure-bridged-accessory",
-    ],
-    optionalCharacteristics: ["name"],
   }],
 ]);